Refactored IndexHolder
